CI fix

+++ src/lpython/python_evaluator.cpp
@@ -145,11 +145,11 @@ Result<PythonCompiler::EvalResult> PythonCompiler::evaluate(
             if (ASRUtils::get_FunctionType(fn)->m_return_var_type->type == ASR::ttypeType::UnsignedInteger) {
                 uint8_t r = e->execfn<uint8_t>(run_fn);
                 result.type = EvalResult::unsignedInteger1;
-                result.u32 = r;
+                result.u8 = r;
             } else {
                 int8_t r = e->execfn<int8_t>(run_fn);
                 result.type = EvalResult::integer1;
-                result.i32 = r;
+                result.i8 = r;
             }
         } else if (return_type == "integer2") {
             ASR::symbol_t *fn = ASR::down_cast<ASR::Module_t>(symbol_table->resolve_symbol(module_name))
@@ -158,11 +158,11 @@ Result<PythonCompiler::EvalResult> PythonCompiler::evaluate(
             if (ASRUtils::get_FunctionType(fn)->m_return_var_type->type == ASR::ttypeType::UnsignedInteger) {
                 uint16_t r = e->execfn<uint16_t>(run_fn);
                 result.type = EvalResult::unsignedInteger2;
-                result.u32 = r;
+                result.u16 = r;
             } else {
                 int16_t r = e->execfn<int16_t>(run_fn);
                 result.type = EvalResult::integer2;
-                result.i32 = r;
+                result.i16 = r;
             }
         } else if (return_type == "integer4") {
             ASR::symbol_t *fn = ASR::down_cast<ASR::Module_t>(symbol_table->resolve_symbol(module_name))
